requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
asgiref==3.8.1
attrs==25.3.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.4
discord.py==2.5.2
Django==5.2.3
django-debug-toolbar==5.2.0
django-environ==0.12.0
django-extensions==4.1
django-filter==25.1
django-rest-framework==0.1.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dotenv==0.9.9
frozenlist==1.7.0
idna==3.10
MarkupSafe==3.0.2
multidict==6.5.1
pillow==11.2.1
propcache==0.3.2
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==25.1.0
PyPDF2==3.0.1
python-dotenv==1.1.1
reportlab==4.4.2
requests==2.32.4
rest-framework-simplejwt==0.0.2
sqlparse==0.5.3
stripe==12.2.0
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
webcolors==24.11.1
Werkzeug==3.1.3
yarl==1.20.1
psycopg2-binary>=2.9
gunicorn
bcrypt
